mm: vmstat: make vmstat_work cpu bound during requeue vmstat_work is currently a per cpu worker thread that requeues itself using schedule_delayed_work().
schedule_delayed_work() makes the worker thread unbound. Since
its unbound, when the timer for the delayed workqueue is migrated,
the current code  path  can cause the per cpu worker to get
executed on the CPU other than what it is intended for causing
undesired effects. This overrides the choice of making the worker
per cpu in the first place.

Fix this by using schedule_delayed_work_on() and make it CPU
bound.